Avoid race-condition in e2e_001_trust_broadcast
Sometimes the 5 seconds sleep in that test before use `try_recv` to
fetch the package was *not* enough.

It also adds an event logger for tests. Each test should call to
`setup()` function at the beginning in order to work even if test is
executed just for that unique test.