Added a flag to specify the required percentage of desired peers we currently have fulfilled.
If we have --baker-min-peer-satisfaction-percentage set to 50(default) and we have a desired amount of peers of value 50, then baking won't start until we have at least 25 peers. Baking doesn't make sense unless we have someone to communicate about it with. The amount of peers is across all networks currently joined to. Maybe it should be changed in the future.

Also a bug in bootstrapping has been fixed where dnssec check flag was negated. I have changed the name of the variable to make it clearer what it does.