Previously if the node had no peers and never got one, then the RPC server would hang forever.
I have changed the method such that if the current peer list doesn't contain the needed peer for the TPS test, then the method will fail with a failed precondition message. That seemed like the most appropriate error type to use for this.

This change means that any client consuming the RPC interface will need to first check the peer list prior to starting a TPS test. There is of course a slight chance that the peer will disappear between the peer list call and the tps test call, but this should not happen very often in the real world.