Update consensus to latest branch of crypto/id.
The main change is in how genesis data is structured, and how it is generated.
At the moment we have an intermediate solution where part of the data is
generated in Haskell and part of it must be provided in files. The reason for
the latter is that we must generate a bunch of keys for identity providers, etc,
and the only facilities for that are in rust code at the moment. After the next
POC we should unify these so that an external tool generates all the necessary
data.

Documentation of the formats will be forthcoming once the parameters settle a
bit more.