Add parameters to generate bakers and make genesis.
We now automatically generate a given number of finalizers so they don't have to
be udpated manually. By default everyone is a finalizer.

We also allow assembing the genesis from multiple files. This is easier to use
when we only want to update some parameters (e.g., just the bakers).