Update the scheduler with new variant of acorn and better linking support.
This is a first step towards counting the linking cost correctly. The idea is
that when deploying a new contract we charge for the size of the linked
term (init + receive + constraint implementations).

The infrastructure now supports this, the only thing left to add is that the
linker keeps track of the sizes of terms.