fix (persistent-treestate): get proper references to existing blocks
When requesting the `blockStatus` associated with a hash, once the
block has been finalized, we were reading the block directly from the
disk on the fly, not checking if we were going to duplicate the
genesis block of the last finalized block and as this decoded block
was not stored anywhere, we instantly were dropping it.

Now in case we are asking for the last finalized block or the genesis
block we will return a proper pointer to that existing block.

This could be improved with some kind of cache for the blocks that
will hold some blocks in memory for some time.