Store transaction outcomes directly in the transaction table.
The hooking mechanism is for the moment disabled, although we may wish to
revisit it in the future.

The current design of outcomes adds minimal overhead on top of the outcomes
stored in block state, but still bakers might not want to store transaction
outcomes directly, and only in the minimal form of hashes.