feature (transactions): replace transaction table with database
Blocks can now store either Transaction or PersistentTransactions
depending if it is the Basic or PersistentImplementation.

Types used in the TreeStateMonad definition have been abstracted to
match the type families declared for each implementation.

The persistent implementation of the TreeStateMonad has two new
databases: transactions and transactionstatus, that store the
transactions in the same way as the old transaction table used to
do. The implementations of the functions related with transactions
has been modified accordingly.

There has been also a huge rework in the way monadic behaviors were
derived. For more information, check the comment at the top of
`GlobalState.hs`.

Some typeclasses were added:
- Convert a b m: implements the ability to convert between types `a`
and `b` in the monad scope. This is used to abstract the ability to
convert between `Transaction`s and `PersistentTransaction`s.
- BlockDataMonad: With the new blocks that have persistent
transactions, some of the data needs monadic actions to be extracted
and this typeclass provides such abstraction.

The format of the serialized blocks is:
- Serialized baked block (containing only transaction hashes instead
of full transactions)
- Serialized block state
- BlockHeight
- Transactioncount
- TransactionsSize
- TransactionEnergyCost