features (transaction): implement persistence for transaction table
For internal changes on globalstate, see
Concordium/consensus/globalstate-mockup@f20473b19e7517ade6e657096e2de0ca52ed6c87.

These changes essentially lift the new behaviors to the `SkovT` monad
and procure conversions when needed to be able to use `Transaction`s
in the scheduler and when serializing blocks for the network.

The derivations for the `SkovT` monad regarding typeclasses defined in
the globalstate package were reworked following the same rules as the
ones in the globalstate package.

The rest are just import fixes as things were moved in the globalstate package.