Update to latest consensus version.
The reason for the u64->i64 change is that all of Haskell's standard library
uses Int for sizes, lengths, etc., so it is most convenient to just use that
directly as opposed to transforming it at some point.

Any negative value has the same meaning as 0; no messages will be sent.