CB-165 fixes
Addresses CB-165.

Remove looking up state when querying block state.
This will speed up queries a bit, but a more substantial changes should be done
so that such queries can be supported.

In particular it would make sense to have a notion of "block summary" that can
be looked up directly and quicly. It should probably be what a block pointer is
now, minus the block data.

This will likely entail revising the block serialization.