Recover finalization state from the initial tree state.
This required some reworking of the abstractions to not duplicate the logic too
much. In particular the SkovQueryMonad is removed as an independent abstraction
and is now just a constraint alias, with methods defined at the top-level.

Since previously most of the methods were trivially implemented either via
BlockStateQuery or TreeStateMonad no generality seems to be lost, however the
change enables an easy implementation of SkovQueryMonad in terms of existing
concrete types.

This is useful because we need to run the computation to reconstruct the initial
finalization state in the context of a SkovQueryMonad.

For this reason we provide the GlobalState type which automatically implements
the SkovQueryMonad and is the type used to reconstruct the initial finalization
state.

The functions `initialFinalizationState` and `initialPassiveFinalizationState`
are no longer used, and are subsumed by `recoverFinalizationState` function.