Slight reworking of how encrypted amounts are stored on the chain.
There are two parts to it, the `selfAmounts` and the queue of incoming
transfers. The `selfAmounts` is the result of account's actions, and thus the
account can fully control it, whereas the incomingTransfers can be added to by
other accounts, by making transfers.

When an encrypted transfer is made the selfAmounts is updated, and it always
factors into the transfer that is being made.