Update consensus for the initial account creation changes.
This also changes the response for the sendTransaction RPC to make use of the boolean. The `false` value will be returned when the submitted transaction is stale, duplicate, or malformed, as opposed to the previous behaviour where it returned an error.

The error is only returned when an internal server error occurs.