Mark concordium-base as a dependency.
It appears that stack has some problems passing correct arguments to the
concordium-base if that is built as a dependency of concordium-consensus,
leading to "impossible just happened" errors.

This seems to be caused due to a combination of template haskell and incorrect
flags passed to `gcc` during code generation. In particular the shared libraries
concordium-base depends on are not being given as arguments.

There are at least the following two ways to fix this
- list concordium-base as a custom dependency
- explicitly pass
```
-L../concordium-base/lib -laggregate_sig -lecvrf leddsa_ed25519 -lencrypted_transfers -lffi_helpers -lid -lsha_2
``` as ghc-options
to the build. This latter is problematic due to relative paths. If we wanted to
go the second route we should set this in the Setup.hs script to avoid problems.

I have chosen the simpler option which seems to work. I don't think there are
significant downsides to this option either.