Fix ban and unban handling to remove crashes and deadlocks.
Resolves CB-1087.
Resolves CB-1050.

The following changes are made
- the flag `--no-trust-bans` has been changed to `--trust-bans`. The default now
is to __not__ trust the ban requests we get from our peers. There is no reason
to trust such bans and they are trivially manufactured, and can thus bring about
banning of a single peer on the entire network.
- ban requests are forwarded only after the node itself has removed the peers,
and only if it has done so.

I do not know the history of the ban mechanism so I did not remove the
retransmission of the messages entirely, although I see little point in it when
it has no authentication or authorization data associated with it. In its
current form it can only lead to abuse.

The reason for the deadlock was the following sequence of calls
- process_network_events acquires a write lock on node.connections and then iterates
over all of them. The lock is held for the duration of the processing of all
connections.
- for each connection (if there are relevant events) a `read_stream` is called
synchronously
- if a complete message is available then process_message is read which
essentially forwards to `handle_incoming_message`
- previusly, if the message was `BanNode` the function `handle_ban` was called
- this called `ban_node` which try to find the connection by either ID or IP,
both of which require acquiring a read lock on node.connections, leading to
either a deadlock or crash
- further if `--no-trust-bans` was __not__ given, ultimately
`send_message_to_all` was called when the ban was supposed to be rebroadcast,
and this tried to acquire a write lock on the connections object. This would either
have deadlocked or crashed the message processing thread, leading the node to
hang. But the node hanged even before this was reached.