Remove the (un)ban network message entirely.
As mentioned in the comments previously, the way it is currently designed,
propagating these messages on the p2p network can only be misused. If we ever
have different tiers of nodes, e.g., a trusted set, and everybody else, then it
might make sense to have these bans propagate on the trusted network.

Additionally move the poll registry socket registration and deregistration into
the ::new and ::drop methods to make it harder to misuse.