Fix a performance issue with account updates.
If an account owned many contract instances any update to it would have been
very slow. This could have been used as a DOS vector.

The simplest thing to do was to remove the accountInstances from the account.
They are not needed at the moment, and it is unclear whether they will ever be
needed. Any alternative solution would have required substantially more changes
at the moment, for no gain.

This does however mean that this is a breaking change since state hashing is
different. Additionally, the node's external API is affected, the getAccountInfo
query does not return the list of instances owned by the account anymore. This
was not used anywhere, so the fallout will be minimal (only the parser in
concordium-client needs to be updated.)

Closes CB-1185