Fix logging components reporting and make the bake loop not skip slots.
Skipping slots has a big negative effect when the load is close to the energy
limit of blocks. It means that the average block times increase and thus the TPS
goes below the expected even though in principle the chain would be able to
handle it.

Add some documentation, and delete unused imports in tests.