Support encrypted baker credentials.
The keys are decrypted early before invoking consensus. The main reason for this
is that we want to avoid using stdin/stderr because that is already used for
logging and is likely to be redirected. I could not find an easy Haskell library
to use.

This adds a new flag --decrypt-baker-credentials to make it explicit when we
are passing encrypted credentials to the baker.