Update rust dependencies to resolve most security advisories.
This MR also fixes sporadic test failure due to race conditions in test
scheduling and overwriting of databases. Part of this was making data and config
dirs mandatory arguments.

This makes the node safer to use when running two of them on a single machine
since it will not silently default to something.