Do not use self-chosen node ids for identifying peers.
Since we cannot trust nodes to choose ids honestly we cannot use chosen ids for
anything.

This adds a `local_id` to each of the remote peers which is used to identify
them. This local id is simply the connection token we already have, which is
guaranteed to be unique.

This prevents a vulnerability where a malicious node could force another node to
drop an arbitrary connection by simply choosing the same node id.