Documentation of the different peer identifiers.
This also fixes and issue where the bootstrapper would serve incorrect peer
ports some of the time. It would return the port on which the initial connection
by the node was initiated, but that is not the port on which the node can be
connected to. I introduced this issue with the remote peer changes in the
previous commits.