Add additional sandboxing for the node process.
The main restriction is to run as a dynamically allocated user and no longer as root.
The effect of this change is that `systemd-analyze security` reports

```
systemd-analyze security concordium-node.service
...
→ Overall exposure level for concordium-node.service: 4.2 OK
```

An upgrade from 5 Medium of before.

Additional sandboxing could be done by limiting system calls, but that is a lot
more tricky, and the benefits are not great for the node.