Fix docker-compose setup (#121)
The 'dev-master' image (intended for use in Docker Compose) is build automatically on a daily basis. The base images are pushed with tag 'latest' by default and only retagged manually (there's currently no process for this).

This PR changes the build job to use the base images with tag 'latest'. Due to the automatic nature of these builds, using 'latest' seems both necessary and natural.

The deployment script has been updated to configure the binaries directly with environment variables as much as possible. The entrypoint scripts are kept on an absolute minimum complexity.

Part of https://concordium.atlassian.net/browse/DO-319.