Retrieve live blocks from the block table as well when trying to get block parents.
This addresses a sporadic node crash in a case when an unfortunate runtime
behaviour combined with the timing of finalization.

Since finalization marks blocks as finalized from most recent backwards it
follows block parent pointers backwards. As a result if the weak pointer is
finalized for some reason before all blocks are finalized it used to crash.

We have not seem many instances of this, but I have just experienced it
consistently during catchup.