Split collecto backend into its own crate.
The node collector and collector backend share a common type that is used to
communicate. Thus the concordium-node now depends on collector-backend.

I did not move the collector itself to the backend package since it has too many
other things in common with the node, such as setting up logging.