Account aliases (#202)
Introduce handling of protocol version 3 account aliases.

The main change is the consolidation of the account map into its own module AccountMap.
This module is used both by the basic and persistent implementations. The basic
implementation previously used a simple Data.Map which is no longer reasonable
when we have to look up a key by a prefix, a task for which a Trie is a perfect
match.

Once this accountmap is introduce the only change in P3 as opposed to P2/P1 is
that for P3 we lookup by a 29-byte account address prefix. The effect of this
will be that you can refer to an account via any address that matches 29 bytes
of the canonical one. The only potential exception to this will be accounts
which clash on the first 29 bytes of the address. This will be disallowed for
new accounts in P3, but even in P1 and P2 there should not be any such accounts
since it involves almost finding a hash collision.