Fix and clean up dockerized builds of node-related components
Probably the biggest change is that genesis data no longer is baked into the node image. Instead, genesis has become it's own first-class component: The data is kept in a separate image that may copy the data into the expected location on startup. The updated documentation includes a full example docker-compose script on how to wire this up.