Simplify contract state operations a bit, reinstate paired state.
The paired implementation is somewhat hacky. It uses basic contract state and
serialization to convert from one type of contract state to another.

This is not ideal, but a more comprehensive solution requires much more
extensive FFI (i.e., paired state constructions, and a paired construction of
state in Rust). This is a lot of complications for something that ultimately
does not matter much. The current solution is useful for basic testing, which is
all that the paired state is used for.