Avoid processing protocol updates during startup if the state is already present.
Concretely this removes the processing of protocol updates for intermediate
protocol updates during startup, as well as caching of the state during startup
for all but the last protocol version.

Additionally we no longer store genesis blocks in-memory. This is significant
since the genesis blocks for protocol updates contain the serialized state. This
significantly decreases memory consumption of the running node.