Remove genesis data from genesis block pointer.
This removes the last place where genesis data was retained by the node.
All those places now use GenesisConfiguration which has a very small footprint.

The only complication was in loading blocks from LMDB due to backwards
compatibility. However the way genesis data is serialized we can load
GenesisConfiguration assuming we know the genesis hash, which we do.

The implementation as is is compatible with existing databases which store the
entire genesis block, as well as with the new way the genesis block is stored.