Boilerplate code for protocol v8 (#1226)
This adds boilerplate code necessary for account version 4, which is part of protocol version 8. This is mainly guided by the compiler and is at the moment exactly the same as account v3, except that the type has been changed. Non-trivial migration functions are implemented as errors and tracked in Github issues.