implementation of `touch`
Adds definitions and implementation of the `touchTokenAccount` operation
in the token kernel and the block state operations.

The interface in the kernel is

```
touch :: PLTAccount m -> m Bool
```

The return value indicates whether the account already had a token
balance associated (False) or if a zero account balance was newly
created (True).