COR-1651: Add `touch` kernel method (#1441)
* implementation of `touch`

Adds definitions and implementation of the `touchTokenAccount` operation
in the token kernel and the block state operations.

The interface in the kernel is

```
touch :: PLTAccount m -> m Bool
```

The return value indicates whether the account already had a token
balance associated (False) or if a zero account balance was newly
created (True).

* `touch` account before `setAccountState`

* extend trace tests

* added changelog entry

* address review

Added a new function `updateTokenAccountState` to Persistent/Account.hs
that deals with updating the token account state table. The function is
then used in Persistent/BlockState.hs abstractions for the persistent
account versions.

* format

* Update concordium-consensus/src/Concordium/GlobalState/Persistent/Account.hs

Co-authored-by: Thomas Dinsdale-Young <td202@users.noreply.github.com>

* move updateTokenAccountState to Account.ProtocolLevelTokens

---------

Co-authored-by: drsk <rsk@concordium.com>
Co-authored-by: Thomas Dinsdale-Young <td202@users.noreply.github.com>